+++ Android/MyApplication/android/widget/annotations.xml
@@ -1,0 +1,5 @@
+<root>
+    <item name='android.widget.Adapter'>
+        <annotation name='java.lang.Deprecated'/>
+    </item>
+</root>
