+++ Android/MyApplication/app/src/main/AndroidManifest.xml
@@ -71,7 +71,7 @@
         -->
         <meta-data
                 android:name="com.google.android.geo.API_KEY"
-                android:value="AIzaSyDtZ0Q1iHHbGLPMj1hJMxo9Cig5jgLw38A"/>
+                android:value="AIzaSyDi9cU1TQ3J-M2sEty4HteQq_ttIaLrJMU"/>
         <meta-data
                 android:name="com.google.android.gms.version"
                 android:value="@integer/google_play_services_version"/>
