+++ Android/MyApplication/app/src/main/java/com/example/openspot/drivewayFragment.kt
@@ -59,12 +59,12 @@ class DrivewayFragment : PreferenceFragmentCompat() {
                                 drivewayButton.setOnPreferenceClickListener {
                                     AuthUI.getInstance()
                                     val intent = Intent(activity!!.baseContext, ListDrivewayActivity::class.java)
-                                    intent.putExtra("Address", a[i - 1])
-                                    intent.putExtra("Latitude", a[i - 2])
-                                    intent.putExtra("Longitude", a[i - 3])
-                                    intent.putExtra("Active", a[i - 4])
+                                    intent.putExtra("Address", a[i - 4])
+                                    intent.putExtra("Latitude", a[i - 3])
+                                    intent.putExtra("Longitude", a[i - 2])
+                                    intent.putExtra("Active", a[i - 1])
                                     intent.putExtra("Price", a[i])
-                                    intent.putExtra("counter", drivewayButton.key)
+                                    intent.putExtra("counter2", drivewayButton.key)
                                     startActivity(intent)
                                     true
                                 }
