+++ Android/MyApplication/app/src/main/res/layout/activity_list_driveway.xml
@@ -72,4 +72,17 @@
             android:layout_marginVertical="8dp"
             android:layout_alignParentRight="true"/>
 
+    <Button
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginVertical="8dp"
+            android:layout_marginHorizontal="8dp"
+            android:layout_alignParentBottom="true"
+            android:layout_alignParentLeft="true"
+            android:text="Delete"
+            android:background="@color/fui_transparent"
+            android:textSize="8pt"
+            android:textColor="#FF0000"
+            android:onClick="deleteButton"/>
+
 </RelativeLayout>
