+++ Android/MyApplication/app/src/main/res/values/strings.xml
@@ -19,5 +19,8 @@
     <string name="fullName">Full Name</string>
     <string name="email">Email</string>
     <string name="DOB">MM DD YYYY</string>
+
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
